+++ src/modules/setup-wizard/setup-wizard.controller.ts
@@ -1,5 +1,5 @@
-import { Body, Controller, Post, UseGuards } from '@nestjs/common';
-import { ApiTags } from '@nestjs/swagger';
+import { Body, Controller, Get, Post, UseGuards } from '@nestjs/common';
+import { ApiOperation, ApiTags } from '@nestjs/swagger';
 
 import { UserDto } from '../users/users.dto';
 import { AuthService } from '../../core/auth/auth.service';
@@ -14,8 +14,20 @@ export class SetupWizardController {
   ) { }
 
   @Post('/create-first-user')
-  async createFirstUser(@Body() body: UserDto) {
+  @ApiOperation({
+    summary: 'Create the first user.',
+    description: 'This endpoint is not available after the Homebridge setup wizard is complete.',
+  })
+  async setupFirstUser(@Body() body: UserDto) {
     return await this.authService.setupFirstUser(body);
   }
 
+  @Get('/get-setup-wizard-token')
+  @ApiOperation({
+    summary: 'Creates a auth token to be used by the setup wizard.',
+    description: 'This endpoint is not available after the Homebridge setup wizard is complete.',
+  })
+  async generateSetupWizardToken() {
+    return await this.authService.generateSetupWizardToken();
+  }
 }