+++ ui/src/app/modules/setup-wizard/setup-wizard.module.ts
@@ -2,11 +2,12 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { TranslateModule } from '@ngx-translate/core';
+import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
 
+import { CoreModule } from '@/app/core/core.module';
 import { SetupWizardRoutingModule } from './setup-wizard-routing.module';
 import { SetupWizardComponent } from './setup-wizard.component';
 import { SetupWizardGuard } from './setup-wizard.guard';
-
 
 @NgModule({
   declarations: [
@@ -17,6 +18,8 @@ import { SetupWizardGuard } from './setup-wizard.guard';
     FormsModule,
     ReactiveFormsModule,
     TranslateModule,
+    NgbModule,
+    CoreModule,
     SetupWizardRoutingModule,
   ],
   providers: [